move subcomponents into their own directory
