+++ src/pages/MinerDashboard/Settings/MinerSettings.modal.tsx
@@ -8,8 +8,8 @@ import { useOpenState } from 'src/hooks/useOpenState';
 import { useActiveCoin } from 'src/rdx/localSettings/localSettings.hooks';
 import { useReduxState } from 'src/rdx/useReduxState';
 import styled from 'styled-components';
-import { NotificationSettings } from './NotificationSettings';
-import { PayoutSettings } from './PayoutSettings';
+import { NotificationSettings } from './components/NotificationSettings/NotificationSettings';
+import { PayoutSettings } from './components/PayoutSettings/PayoutSettings';
 
 const Split = styled.div`
   display: flex;
