+++ src/pages/MinerDashboard/Settings/components/NotificationSettings/NotificationSettings.tsx
